# [SPARK-53075][CORE][TESTS] Use Java `Files.readAllLines/write` instead of `FileUtils.(read|write)Lines`