--- a/common/utils/pom.xml
+++ b/common/utils/pom.xml
@@ -52,11 +52,6 @@
       <artifactId>jackson-module-scala_${scala.binary.version}</artifactId>
     </dependency>
     <dependency>
-      <groupId>commons-io</groupId>
-      <artifactId>commons-io</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
       <groupId>org.apache.ivy</groupId>
       <artifactId>ivy</artifactId>
     </dependency>