@LuciferYang 2025-08-02
Merged into master for Apache Spark 4.1.0. Thanks @dongjoon-hyun and @peter-toth